+++ src/Oro/Bundle/PlatformBundle/EventListener/Console/OptionalListenersListener.php
@@ -2,12 +2,8 @@
 
 namespace Oro\Bundle\PlatformBundle\EventListener\Console;
 
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Event\ConsoleCommandEvent;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
-
-use Oro\Bundle\PlatformBundle\Command\OptionalListenersCommand;
 use Oro\Bundle\PlatformBundle\Manager\OptionalListenerManager;
 
 class OptionalListenersListener
@@ -39,34 +35,6 @@ class OptionalListenersListener
         if (!empty($listeners)) {
             $this->listenersManager->disableListeners($listeners);
         }
-    }
-
-    /**
-     * @deprecated since 2.4,
-     * should be called in Oro\Bundle\PlatformBundle\EventListener\Console\OptionalListenersOptionsListener
-     *
-     * @param Command $command
-     * @param InputInterface $input
-     */
-    protected function addOptionsToCommand(Command $command, InputInterface $input)
-    {
-        $inputDefinition = $command->getApplication()->getDefinition();
-        $disableOptionalListenerOption = new InputOption(
-            self::DISABLE_OPTIONAL_LISTENERS,
-            null,
-            InputOption::VALUE_REQUIRED|InputOption::VALUE_IS_ARRAY,
-            sprintf(
-                'Disable optional listeners, "%s" to disable all listeners, '
-                .'command "%s" shows all listeners',
-                self::ALL_OPTIONAL_LISTENERS_VALUE,
-                OptionalListenersCommand::NAME
-            )
-        );
-        /**
-         * Starting from Symfony 2.8 event 'ConsoleCommandEvent' fires after all definitions were merged.
-         */
-        $inputDefinition->addOption($disableOptionalListenerOption);
-        $command->getDefinition()->addOption($disableOptionalListenerOption);
     }
 
     /**